+++ utility_functions.h
@@ -10,6 +10,6 @@ vector<int> choose_percent(int n, double percent);
 vector<int> choose_neighbours(int n, int k, vector<int> excluded);
 bool check_connected(vector<vector<int>>& al);
 int uniform_distribution(int min, int max);
-double exponential_distribution(double mean);
+int exponential_distribution(long long mean);
 
 #endif //UTILITY_H
